gateway/requirements.txt
fastapi
uvicorn
psutil
pydantic
paho-mqtt
